tests/wpt/tests/websockets/interfaces/WebSocket/bufferedAmount/bufferedAmount-defineProperty-setter.html

<!doctype html>
<meta charset=utf-8>
<title>WebSockets: defineProperty setter for bufferedAmount</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function() {
  window.setter_ran = false;
  Object.defineProperty(WebSocket.prototype, 'bufferedAmount', {
    set: function(v) { window[v] = true; }
  });
  var ws = new WebSocket('ws://example.invalid/');
  ws.bufferedAmount = 'setter_ran';
  assert_true(setter_ran);
});
</script>
